Move Group.ts to src/Query/
