+++ tests/Group.test.ts
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 import moment from 'moment';
-import { Group } from '../src/Group';
+import { Group } from '../src/Query/Group';
 import type { Grouping, GroupingProperty } from '../src/Query';
 import type { Task } from '../src/Task';
 import { fromLine } from './TestHelpers';
